pyproject.toml
[project]
name = "sg-jobs"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.28.0",
    "pdfplumber>=0.11.0",
    "beautifulsoup4>=4.14.0",
    "rapidfuzz>=3.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.2.0",
]
